better event handling

+++ src/relations/db.py
@@ -134,9 +134,6 @@ class DbProvides(Object):
         If the backend relation is fully initialised and available, we generate the proposed
         database and create a user on the postgres charm, and add preliminary data to the databag.
         """
-        if not self.charm.unit.is_leader():
-            return
-
         if not self.charm.backend.postgres:
             # We can't relate an app to the backend database without a backend postgres relation
             wait_str = "waiting for backend-database relation to connect"
@@ -178,6 +175,18 @@ class DbProvides(Object):
             join_event.defer()
             return
 
+        self.update_databag(
+            join_event.relation,
+            {
+                "user": user,
+                "password": password,
+                "database": database,
+            },
+        )
+
+        if not self.charm.unit.is_leader():
+            return
+
         user = self._generate_username(join_event)
         password = pgb.generate_password()
 
@@ -207,15 +216,6 @@ class DbProvides(Object):
         cfg.add_user(user, admin=self.admin)
         self.charm.render_pgb_config(cfg, reload_pgbouncer=True)
 
-        self.update_databag(
-            join_event.relation,
-            {
-                "user": user,
-                "password": password,
-                "database": database,
-            },
-        )
-
     def _on_relation_changed(self, change_event: RelationChangedEvent):
         """Handle db-relation-changed event.
 
@@ -237,13 +237,13 @@ class DbProvides(Object):
             f"DEPRECATION WARNING - {self.relation_name} is a legacy relation, and will be deprecated in a future release. "
         )
 
-        # No backup values because if pgb_app_databag isn't populated, this relation isn't
-        # initialised. This means that the database and user requested in this relation haven't
-        # been created, so we defer this event until the databag is populated.
-        pgb_app_databag = change_event.relation.data[self.charm.app]
-        database = pgb_app_databag.get("database")
-        user = pgb_app_databag.get("user")
-        password = pgb_app_databag.get("password")
+        # No backup values because if databag isn't populated, this relation isn't initialised.
+        # This means that the database and user requested in this relation haven't been created,
+        # so we defer this event until the databag is populated.
+        databag = self.get_databags(change_event.relation)[0]
+        database = databag.get("database")
+        user = databag.get("user")
+        password = databag.get("password")
 
         if None in [database, user, password]:
             logger.warning(
@@ -270,10 +270,10 @@ class DbProvides(Object):
 
     def update_port(self, relation: Relation, port: str):
         """Updates databag to match new port."""
-        pgb_app_databag = relation.data[self.charm.app]
-        database = pgb_app_databag.get("database")
-        user = pgb_app_databag.get("user")
-        password = pgb_app_databag.get("password")
+        databag = self.get_databags(relation)[0]
+        database = databag.get("database")
+        user = databag.get("user")
+        password = databag.get("password")
 
         if None in [database, user, password]:
             logger.warning("relation not fully initialised - skipping port update")
@@ -300,14 +300,14 @@ class DbProvides(Object):
 
     def update_postgres_endpoints(self, relation: Relation, reload_pgbouncer: bool = False):
         """Updates postgres replicas."""
-        database = relation.data[self.charm.app].get("database")
+        database = self.get_databags(relation)[0].get("database")
         if database is None:
             logger.warning("relation not fully initialised - skipping postgres endpoint update")
             return
 
         # In postgres, "endpoints" will only ever have one value. Other databases using the library
         # can have more, but that's not planned for the postgres charm.
-        postgres_endpoint = self.charm.backend.app_databag.get("endpoints")
+        postgres_endpoint = self.charm.backend.postgres_databag.get("endpoints")
 
         cfg = self.charm.read_pgb_config()
         cfg["databases"][database] = {
@@ -330,51 +330,6 @@ class DbProvides(Object):
         # Write config data to charm filesystem
         self.charm.render_pgb_config(cfg, reload_pgbouncer=reload_pgbouncer)
 
-    def update_databag(self, relation, updates: Dict[str, str]):
-        """Updates databag with the given dict."""
-        databags = [relation.data[self.charm.unit]]
-        if self.charm.unit.is_leader():
-            databags.append(relation.data[self.charm.app])
-
-        # Databag entries can only be strings
-        for key, item in updates.items():
-            updates[key] = str(item)
-
-        for databag in databags:
-            databag.update(updates)
-
-    def _generate_username(self, event):
-        """Generates a unique username for this relation."""
-        app_name = self.charm.app.name
-        relation_id = event.relation.id
-        model_name = self.model.name
-        return f"{app_name}_user_id_{relation_id}_{model_name}".replace("-", "_")
-
-    def _get_read_only_endpoint(self):
-        """Get a read-only-endpoint from backend relation.
-
-        Though multiple readonly endpoints can be provided by the new backend relation, only one
-        can be consumed by this legacy relation.
-        """
-        read_only_endpoints = self.charm.backend.app_databag.get("read-only-endpoints")
-        if read_only_endpoints is None or len(read_only_endpoints) == 0:
-            return None
-        return read_only_endpoints.split(",")[0]
-
-    def _get_state(self) -> str:
-        """Gets the given state for this unit.
-
-        Args:
-            standbys: the comma-separated list of postgres standbys
-
-        Returns:
-            The described state of this unit. Can be 'standalone', 'master', or 'standby'.
-        """
-        if self.charm.unit.is_leader():
-            return "master"
-        else:
-            return "standby"
-
     def _on_relation_departed(self, departed_event: RelationDepartedEvent):
         """Handle db-relation-departed event.
 
@@ -386,11 +341,9 @@ class DbProvides(Object):
             f"DEPRECATION WARNING - {self.relation_name} is a legacy relation, and will be deprecated in a future release. "
         )
 
-        app_databag = departed_event.relation.data[self.charm.app]
-        unit_databag = departed_event.relation.data[self.charm.unit]
-
-        for databag in [app_databag, unit_databag]:
-            databag["allowed-units"] = self.get_allowed_units(departed_event.relation)
+        self.update_databag({
+            "allowed-units": self.get_allowed_units(departed_event.relation)
+        })
 
     def _on_relation_broken(self, broken_event: RelationBrokenEvent):
         """Handle db-relation-broken event.
@@ -401,9 +354,9 @@ class DbProvides(Object):
         This doesn't delete any tables so we aren't deleting a user's entire database with one
         command.
         """
-        app_databag = broken_event.relation.data[self.charm.app]
-        user = app_databag.get("user")
-        database = app_databag.get("database")
+        databag = self.get_databags(broken_event.relation)[0]
+        user = databag.get("user")
+        database = databag.get("database")
 
         if not self.charm.backend.postgres or None in [user, database]:
             # this relation was never created, so wait for it to be initialised before removing
@@ -434,7 +387,56 @@ class DbProvides(Object):
 
         cfg.remove_user(user)
         self.charm.render_pgb_config(cfg, reload_pgbouncer=True)
-        self.charm.backend.postgres.delete_user(user)
+        if self.charm.unit.is_leader():
+            self.charm.backend.postgres.delete_user(user)
+
+    def update_databag(self, relation, updates: Dict[str, str]):
+        """Updates databag with the given dict."""
+        # Databag entries can only be strings
+        for key, item in updates.items():
+            updates[key] = str(item)
+
+        for databag in self.get_databags(relation):
+            databag.update(updates)
+
+    def _generate_username(self, event):
+        """Generates a unique username for this relation."""
+        app_name = self.charm.app.name
+        relation_id = event.relation.id
+        model_name = self.model.name
+        return f"{app_name}_user_id_{relation_id}_{model_name}".replace("-", "_")
+
+    def get_databags(self, relation):
+        """Returns available databags for the given relation"""
+        databags = [relation.data[self.charm.unit]]
+        if self.charm.unit.is_leader():
+            databags.append(relation.data[self.charm.app])
+        return databags
+
+    def _get_read_only_endpoint(self):
+        """Get a read-only-endpoint from backend relation.
+
+        Though multiple readonly endpoints can be provided by the new backend relation, only one
+        can be consumed by this legacy relation.
+        """
+        read_only_endpoints = self.charm.backend.postgres_databag.get("read-only-endpoints")
+        if read_only_endpoints is None or len(read_only_endpoints) == 0:
+            return None
+        return read_only_endpoints.split(",")[0]
+
+    def _get_state(self) -> str:
+        """Gets the given state for this unit.
+
+        Args:
+            standbys: the comma-separated list of postgres standbys
+
+        Returns:
+            The described state of this unit. Can be 'standalone', 'master', or 'standby'.
+        """
+        if self.charm.unit.is_leader():
+            return "master"
+        else:
+            return "standby"
 
     def get_allowed_subnets(self, relation: Relation) -> str:
         """Gets the allowed subnets from this relation."""
